Added several translations (not translated yet), enhanced build scripts, fixed minor bugs on compilation

+++ trunk/SQLiteStudio3/create_linux_portable.sh
@@ -9,7 +9,7 @@ if [ "$#" -lt 2 ] || [ "$#" -gt 3 ]; then
   exit 1
 fi
 
-if [ "$#" -eq 3 ] && [ "$3" != "tgz" ] && [ "$3" != "dist" ]; then
+if [ "$#" -eq 3 ] && [ "$3" != "tgz" ] && [ "$3" != "dist" ] && [ "$3" != "dist_plugins" ]; then
   printUsage
   exit 1
 fi
@@ -106,31 +106,33 @@ VERSION=`SQLiteStudio/sqlitestudiocli -v | awk '{print $2}'`
 if [ "$3" == "tgz" ]; then
   tar cf sqlitestudio-$VERSION.tar SQLiteStudio
   xz -z sqlitestudio-$VERSION.tar
-elif [ "$3" == "dist" ]; then
-  # Complete
-  echo "Building complete package: sqlitestudio-$VERSION.tar.xz"
-  tar cf sqlitestudio-$VERSION.tar SQLiteStudio
-  xz -z sqlitestudio-$VERSION.tar
+elif [ "$3" == "dist" ] || [ "$3" == "dist_plugins" ]; then
+  if [ "$3" == "dist" ]; then
+    # Complete
+    echo "Building complete package: sqlitestudio-$VERSION.tar.xz"
+    tar cf sqlitestudio-$VERSION.tar SQLiteStudio
+    xz -z sqlitestudio-$VERSION.tar
   
-  # App
-  echo "Building incremental update package: sqlitestudio-$VERSION.tar.gz"
-  cp -R SQLiteStudio app
-  cd app
-  rm -rf plugins
-  rm -f lib/libQ*
-  rm -rf iconengines
-  rm -rf imageformats
-  rm -rf platforms
-  rm -rf platformthemes
-  rm -rf printsupport
-  rm -f lib/libicu*
-  rm -f lib/libsqlite.so.0 ;# this is for SQLite 2
-  find . -type l -exec rm -f {} \;
-  tar cf sqlitestudio-$VERSION.tar *
-  gzip -9 sqlitestudio-$VERSION.tar
-  mv sqlitestudio-$VERSION.tar.gz ..
-  cd ..
-  rm -rf app
+    # App
+    echo "Building incremental update package: sqlitestudio-$VERSION.tar.gz"
+    cp -R SQLiteStudio app
+    cd app
+    rm -rf plugins
+    rm -f lib/libQ*
+    rm -rf iconengines
+    rm -rf imageformats
+    rm -rf platforms
+    rm -rf platformthemes
+    rm -rf printsupport
+    rm -f lib/libicu*
+    rm -f lib/libsqlite.so.0 ;# this is for SQLite 2
+    find . -type l -exec rm -f {} \;
+    tar cf sqlitestudio-$VERSION.tar *
+    gzip -9 sqlitestudio-$VERSION.tar
+    mv sqlitestudio-$VERSION.tar.gz ..
+    cd ..
+    rm -rf app
+  fi
 
   # Plugins
   mkdir plugins
